Fix duplicated snackbar issue
