+++ app/src/main/java/com/github/k409/fitflow/ui/FitFlowApp.kt
@@ -4,7 +4,6 @@ package com.github.k409.fitflow.ui
 
 import android.icu.text.CompactDecimalFormat
 import androidx.compose.animation.core.tween
-import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.Arrangement
@@ -12,7 +11,6 @@ import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.imePadding
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.windowInsetsPadding
@@ -20,11 +18,9 @@ import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.ArrowBack
-import androidx.compose.material.icons.filled.Close
 import androidx.compose.material.icons.outlined.PersonOutline
 import androidx.compose.material3.AssistChipDefaults
 import androidx.compose.material3.BottomAppBarDefaults
-import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.ElevatedAssistChip
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
@@ -33,23 +29,14 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.NavigationBar
 import androidx.compose.material3.NavigationBarItem
 import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Snackbar
-import androidx.compose.material3.SnackbarData
-import androidx.compose.material3.SnackbarHost
 import androidx.compose.material3.SnackbarHostState
 import androidx.compose.material3.Surface
-import androidx.compose.material3.SwipeToDismissBox
-import androidx.compose.material3.SwipeToDismissBoxState
-import androidx.compose.material3.SwipeToDismissBoxValue
 import androidx.compose.material3.Text
-import androidx.compose.material3.TextButton
 import androidx.compose.material3.TopAppBar
 import androidx.compose.material3.TopAppBarDefaults
-import androidx.compose.material3.rememberSwipeToDismissBoxState
 import androidx.compose.material3.surfaceColorAtElevation
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
-import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.MutableState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
@@ -59,7 +46,6 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.Shape
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.painterResource
@@ -79,6 +65,7 @@ import com.github.k409.fitflow.R
 import com.github.k409.fitflow.model.User
 import com.github.k409.fitflow.model.isProfileComplete
 import com.github.k409.fitflow.ui.common.LocalSnackbarHostState
+import com.github.k409.fitflow.ui.common.SwipeableSnackbar
 import com.github.k409.fitflow.ui.navigation.FitFlowNavGraph
 import com.github.k409.fitflow.ui.navigation.NavRoutes
 import java.util.Locale
@@ -169,106 +156,6 @@ fun FitFlowApp(
             )
         }
     }
-}
-
-
-@Composable
-fun SwipeableSnackbar(
-    snackbarHostState: SnackbarHostState,
-    modifier: Modifier = Modifier,
-    dismissSnackbarState: SwipeToDismissBoxState = rememberSwipeToDismissBoxState(
-        confirmValueChange = { value ->
-            if (value != SwipeToDismissBoxValue.Settled) {
-                snackbarHostState.currentSnackbarData?.dismiss()
-                false
-            } else {
-                false
-            }
-        }),
-) {
-    LaunchedEffect(dismissSnackbarState.currentValue) {
-        if (dismissSnackbarState.currentValue != SwipeToDismissBoxValue.Settled) {
-            dismissSnackbarState.reset()
-        }
-    }
-
-    SwipeToDismissBox(
-        modifier = modifier,
-        state = dismissSnackbarState,
-        backgroundContent = {},
-        content = {
-            SnackbarHost(
-                hostState = snackbarHostState,
-                modifier = Modifier.imePadding(),
-                snackbar = {
-                    FitFlowSnackbar(
-                        snackbarData = it,
-                    )
-                },
-            )
-        },
-    )
-}
-
-@Composable
-fun FitFlowSnackbar(
-    snackbarData: SnackbarData,
-    modifier: Modifier = Modifier,
-    actionOnNewLine: Boolean = false,
-    borderStroke: BorderStroke = BorderStroke(
-        width = 1.0.dp,
-        color = MaterialTheme.colorScheme.outline,
-    ),
-    shape: Shape = MaterialTheme.shapes.medium,
-    containerColor: Color = MaterialTheme.colorScheme.secondaryContainer,
-    contentColor: Color = MaterialTheme.colorScheme.onSecondaryContainer,
-    actionColor: Color = MaterialTheme.colorScheme.onSecondaryContainer,
-    actionContentColor: Color = MaterialTheme.colorScheme.onSecondaryContainer,
-    dismissActionContentColor: Color = MaterialTheme.colorScheme.onSecondaryContainer,
-) {
-    val actionLabel = snackbarData.visuals.actionLabel
-    val actionComposable: (@Composable () -> Unit)? = if (actionLabel != null) {
-        @Composable {
-            TextButton(
-                colors = ButtonDefaults.textButtonColors(contentColor = actionColor),
-                onClick = { snackbarData.performAction() },
-                content = { Text(actionLabel) },
-            )
-        }
-    } else {
-        null
-    }
-    val dismissActionComposable: (@Composable () -> Unit)? =
-        if (snackbarData.visuals.withDismissAction) {
-            @Composable {
-                IconButton(
-                    onClick = { snackbarData.dismiss() },
-                    content = {
-                        Icon(
-                            Icons.Filled.Close,
-                            contentDescription = stringResource(R.string.dismiss_snackbar),
-                        )
-                    },
-                )
-            }
-        } else {
-            null
-        }
-
-    Snackbar(
-        modifier = modifier
-            .padding(12.dp)
-            .border(borderStroke, shape),
-        action = actionComposable,
-        dismissAction = dismissActionComposable,
-        actionOnNewLine = actionOnNewLine,
-        shape = shape,
-        containerColor = containerColor,
-        contentColor = contentColor,
-        actionContentColor = actionContentColor,
-        dismissActionContentColor = dismissActionContentColor,
-        content = { Text(snackbarData.visuals.message) },
-    )
 }
 
 @Composable
