+++ app/src/main/java/com/github/k409/fitflow/ui/common/Snackbar.kt
@@ -1,0 +1,150 @@
+@file:OptIn(ExperimentalMaterial3Api::class)
+
+package com.github.k409.fitflow.ui.common
+
+import androidx.compose.foundation.BorderStroke
+import androidx.compose.foundation.border
+import androidx.compose.foundation.layout.imePadding
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Close
+import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Snackbar
+import androidx.compose.material3.SnackbarData
+import androidx.compose.material3.SnackbarDuration
+import androidx.compose.material3.SnackbarHost
+import androidx.compose.material3.SnackbarHostState
+import androidx.compose.material3.SwipeToDismissBox
+import androidx.compose.material3.SwipeToDismissBoxState
+import androidx.compose.material3.SwipeToDismissBoxValue
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
+import androidx.compose.material3.rememberSwipeToDismissBoxState
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.Shape
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.dp
+import com.github.k409.fitflow.R
+
+@Composable
+fun SwipeableSnackbar(
+    snackbarHostState: SnackbarHostState,
+    modifier: Modifier = Modifier,
+    dismissSnackbarState: SwipeToDismissBoxState = rememberSwipeToDismissBoxState(
+        confirmValueChange = { value ->
+            if (value != SwipeToDismissBoxValue.Settled) {
+                snackbarHostState.currentSnackbarData?.dismiss()
+                true
+            } else {
+                false
+            }
+        }),
+) {
+    LaunchedEffect(dismissSnackbarState.currentValue) {
+        if (dismissSnackbarState.currentValue != SwipeToDismissBoxValue.Settled) {
+            dismissSnackbarState.reset()
+        }
+    }
+
+    SwipeToDismissBox(
+        modifier = modifier,
+        state = dismissSnackbarState,
+        backgroundContent = {},
+        content = {
+            SnackbarHost(
+                hostState = snackbarHostState,
+                modifier = Modifier.imePadding(),
+                snackbar = {
+                    FitFlowSnackbar(
+                        snackbarData = it,
+                    )
+                },
+            )
+        },
+    )
+}
+
+@Composable
+fun FitFlowSnackbar(
+    snackbarData: SnackbarData,
+    modifier: Modifier = Modifier,
+    actionOnNewLine: Boolean = false,
+    borderStroke: BorderStroke = BorderStroke(
+        width = 1.0.dp,
+        color = MaterialTheme.colorScheme.outline,
+    ),
+    shape: Shape = MaterialTheme.shapes.medium,
+    containerColor: Color = MaterialTheme.colorScheme.secondaryContainer,
+    contentColor: Color = MaterialTheme.colorScheme.onSecondaryContainer,
+    actionColor: Color = MaterialTheme.colorScheme.onSecondaryContainer,
+    actionContentColor: Color = MaterialTheme.colorScheme.onSecondaryContainer,
+    dismissActionContentColor: Color = MaterialTheme.colorScheme.onSecondaryContainer,
+) {
+    val actionLabel = snackbarData.visuals.actionLabel
+    val actionComposable: (@Composable () -> Unit)? = if (actionLabel != null) {
+        @Composable {
+            TextButton(
+                colors = ButtonDefaults.textButtonColors(contentColor = actionColor),
+                onClick = { snackbarData.performAction() },
+                content = { Text(actionLabel) },
+            )
+        }
+    } else {
+        null
+    }
+    val dismissActionComposable: (@Composable () -> Unit)? =
+        if (snackbarData.visuals.withDismissAction) {
+            @Composable {
+                IconButton(
+                    onClick = { snackbarData.dismiss() },
+                    content = {
+                        Icon(
+                            Icons.Filled.Close,
+                            contentDescription = stringResource(R.string.dismiss_snackbar),
+                        )
+                    },
+                )
+            }
+        } else {
+            null
+        }
+
+    Snackbar(
+        modifier = modifier
+            .padding(12.dp)
+            .border(borderStroke, shape),
+        action = actionComposable,
+        dismissAction = dismissActionComposable,
+        actionOnNewLine = actionOnNewLine,
+        shape = shape,
+        containerColor = containerColor,
+        contentColor = contentColor,
+        actionContentColor = actionContentColor,
+        dismissActionContentColor = dismissActionContentColor,
+        content = { Text(snackbarData.visuals.message) },
+    )
+}
+
+suspend fun SnackbarHostState.showSnackbarIfNotVisible(
+    message: String,
+    actionLabel: String? = null,
+    withDismissAction: Boolean = false,
+    duration: SnackbarDuration =
+        if (actionLabel == null) SnackbarDuration.Short else SnackbarDuration.Indefinite
+) {
+    if (this.currentSnackbarData?.visuals?.message != message) {
+        this.showSnackbar(
+            message = message,
+            actionLabel = actionLabel,
+            withDismissAction = withDismissAction,
+            duration = duration,
+        )
+    }
+}
